% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformfunction.R
\name{tool}
\alias{tool}
\alias{pu}
\alias{pu_inv}
\alias{normalize}
\alias{logitinv}
\alias{my_logit}
\alias{my_surv}
\alias{prev_fit_cste}
\title{tool functions}
\usage{
pu(x, beta)

pu_inv(x, beta, u)

normalize(x)

logitinv(x)

my_logit(x, y, off = NULL, beta = NULL, lam = 0, pen = 2)

my_surv(x, y, atrisk, off = NULL, beta = NULL, lam = 0)

prev_fit_cste(u1, u2, fit)
}
\arguments{
\item{x}{numeric vector or matrix.}

\item{y, u, u1, u2, beta, atrisk}{numeric vector.}

\item{off}{numeric value, offset.}

\item{lam}{numeric value, penalty parameter.}

\item{pen}{hyper-parameter that used in MCP and SCAD penalty functions.}

\item{fit}{a S3 class of cste.}
}
\description{
tool functions
}
\keyword{internal}
